Reorganized for easy installation with stow and added some files. Also added an install script
