+++ bash/.bash_aliases
