+++ bash/.bash_logout
@@ -1,0 +1,3 @@
+#
+# ~/.bash_logout
+#
