+++ conky/.conkyrc
@@ -12,12 +12,12 @@ use_xft			yes
 xftfont			Source Code Pro:size=5
 override_utf8_locale	yes
 minimum_size		1366 0
-maximum_width		1178
+maximum_width		1184
 use_spacer		none
 gap_x			91
 gap_y			58
 
 TEXT
-${execpi 180 getweather -d -s ' ' -t}\
+${execpi 600 getweather -f '%D %tC'}\
 ${alignc -25}${time %a %b %d %I:%M:%S %p}\
 ${alignr}${battery_percent BAT}% ${battery_bar 5,16 BAT}
