+++ firefox/.vimperatorrc
@@ -1,0 +1,1 @@
+colorscheme twily
