+++ install
@@ -1,0 +1,6 @@
+#!/bin/bash
+folders=("bash" "conky" "firefox" "i3" "ranger" "vim" "x")
+for f in "${folders[@]}"
+do
+	stow --verbose=2 -t ~ $f
+done
