+++ ranger/.config/ranger/rc.conf
@@ -1,7 +1,7 @@
 ##################
 ### APPEARANCE ###
 ##################
-set colorscheme classytouch
+set colorscheme colorscheme
 set column_ratios 1,1
 set dirname_in_tabs false
 set display_size_in_main_column true
