+++ ranger/.config/ranger/rifle.conf
