+++ ranger/.config/ranger/scope.sh
