+++ scripts/getweather
@@ -8,59 +8,70 @@ import argparse
 # Parse all the arguments
 parser = argparse.ArgumentParser()
 parser.add_argument(
+    "-f", "--format",
+    nargs='?',
+    help="Format string for output")
+parser.add_argument(
     "-l", "--location",
     nargs='?',
-    help="Location of weather data (default is Waterloo Ontario Canada)")
-parser.add_argument(
-    "-s", "--separator",
-    nargs='?',
-    help="Separator between information (default is ' ')")
+    help="Location of weather data. Default is 'Waterloo Ontario Canada'.")
 parser.add_argument(
     "-i", "--imperial",
-    help="Set measurement system for all units to imperial (default is metric)",
-    action="store_true")
-parser.add_argument(
-    "-d", "--detailed",
-    help="Shows a more detailed description",
-    action="store_true")
-parser.add_argument(
-    "-w", "--windspeed",
-    help="Show wind data in output",
-    action="store_true")
-parser.add_argument(
-    "-t", "--temperature",
-    help="Show temperature data in output",
+    help="Set measurement system for all units to imperial. Default is metric.",
     action="store_true")
 args = parser.parse_args()
 
 # Set the default values if none were provided
 if str(args.location) == "None":
     args.location = "Waterloo Ontario Canada"
-if str(args.separator) == "None":
-    args.separator = " "
+if str(args.format) == "None":
+    args.format = "%D %tC"
 
 # Get the json data from OpenWeatherMap
 url = "http://api.openweathermap.org/data/2.5/weather?q="+str(args.location)
 data = requests.get(url).content
 decoded = json.loads(data)
 
-# Form the output string
-if args.detailed:
-    output = decoded['weather'][0]['description'].capitalize()
-else:
-    output = decoded['weather'][0]['main']
-if args.temperature:
-    temp = decoded['main']['temp'] - 273.15
-    if args.imperial:
-        temp = temp * 1.8 + 32
-        output += args.separator+str(round(temp,1))+'F'
-    else:
-        output += args.separator+str(round(temp,1))+'C'
-if args.windspeed:
-    windspeed = decoded['wind']['speed'] * 3600 / 1000
-    if args.imperial:
-        windspeed = windspeed * 1.61
-        output += args.separator+str(round(windspeed,1))+'mph'
-    else:
-        output += args.separator+str(round(windspeed,1))+'kph'
+# Define all the relevant information
+detailed_description = decoded['weather'][0]['description'].capitalize()
+description = decoded['weather'][0]['main']
+windspeed = str(round(decoded['wind']['speed'] * 3600 / 1000,1))
+temperature = str(round(decoded['main']['temp'] - 273.15,1))
+if args.imperial:
+    temperature = temperature * 1.8 + 32
+    windspeed = windspeed * 1.61
+icon = decoded['weather'][0]['icon']
+
+# Get the UTF8 icon
+if icon == "01d":
+    icon = u"☀"
+elif icon == "01n":
+    icon = u"★"
+elif icon == "02d" or icon == "02n":
+    icon = u"☁"
+elif icon == "03d" or icon == "03n":
+    icon = u"☁"
+elif icon == "04d" or icon == "04n":
+    icon = u"☁"
+elif icon == "09d" or icon == "09n":
+    icon = u"☂"
+elif icon == "10d" or icon == "10n":
+    icon = u"☔"
+elif icon == "11d" or icon == "11n":
+    icon = u"⚡"
+elif icon == "13d" or icon == "13n":
+    icon = u"❄"
+elif icon == "50d" or icon == "50n":
+    icon = u"♒"
+# Special case for extreme weather
+if str(decoded['weather'][0]['id']).startswith('9'):
+    icon = "!"
+
+# Format and print the output
+output = args.format
+output = output.replace("%i", icon)
+output = output.replace("%d", description)
+output = output.replace("%D", detailed_description)
+output = output.replace("%t", temperature)
+output = output.replace("%w", windspeed)
 print(output)
