+++ scripts/i3-exit
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-res=$(dmenu -sb "#af652f" -x 600 -y 320 -w 150 -h 30 -l 3 -p ">" < ~/.dmenu-i3exit)
+res=$(rofi -dmenu < ~/.dmenu-i3exit)
 
 if [ $res = "logout" ]; then
 	i3-msg exit
@@ -11,3 +11,4 @@ fi
 if [ $res = "shutdown" ]; then
 	poweroff
 fi
+exit 0
