+++ vim/.vim/.netrwhist
@@ -1,0 +1,5 @@
+let g:netrw_dirhistmax  =10
+let g:netrw_dirhist_cnt =3
+let g:netrw_dirhist_1='/home/cjbarrac/.i3'
+let g:netrw_dirhist_2='/home/cjbarrac/.ncmpcpp'
+let g:netrw_dirhist_3='/home/cjbarrac/.vim'
