+++ x/.Xresources
@@ -52,3 +52,13 @@ urxvt*transparent:		false
 urxvt*url-select.launcher:	chromium
 urxvt*url-select.underline:	true
 urxvt*cursorBlink:		false
+
+!!! ROFI !!!
+rofi.background: #1a1a1a
+rofi.foreground: #857d59
+rofi.highlightbg: #857d59
+rofi.highlightfg: #1a1a1a
+rofi.bordercolor: #1a1a1a
+rofi.font: Source Code Pro 8
+rofi.padding: 3
+rofi.borderwidth: 0
