+++ x/.xinitrc
@@ -1,0 +1,1 @@
+xbindkeys
